All compiling but still 1 & 6 test failures (when inc. the "it" tests)
Duplicated many of the test sources and resources. Need to clean up later.
