+++ cats/src/it/resources/examples/partnerships/partnerships.csv
@@ -1,0 +1,38 @@
+Id,First,Last,
+1,Adam,Sullivan,
+2,Amy,Avagadro,
+3,Ann,Peterson,
+4,Barb,Goldman,
+5,Barbara,Gilbert,
+6,Bernice,Field,
+7,Bob,Willmot,
+8,Brian,Hill,
+9,Catherine,Gerson,
+10,Christina,Murphy,
+11,Dan,Majewski,
+12,Dana,Schwartz,
+13,Fern,Lobo,
+14,Florine,Schoen,
+15,Gwen,Salmon,
+16,Irene,Zeh,
+17,Isabelle,Engelsted,
+18,Jeanne,Burton,
+19,Judi,Majewski,
+20,Judy,Weinstein,
+21,Kathy,Johnson,
+22,Lanny,Starr,
+23,Marcia,Robinson,
+24,Marian,Levinstein,
+25,Mary,Scerbo,
+26,Mary,Piltch,
+27,Maureen,Harrington,
+28,Nancy,Olt,
+29,Patty,Grasso,
+30,Paul ,Murphy,
+31,Rebecca,Schwartz,
+32,Becky,Singer,
+33,Sally,Kessel,
+34,Sharon,Harowitz,
+35,Sheila,Selby,
+36,Steve ,Bloom,
+37,Susie,Messinger,
