+++ cats/src/it/resources/examples/teamproject/TeamProject.csv
@@ -1,0 +1,7 @@
+Team Number,Team Member 1,Team Member 2,Team Member 3,Team Member 4,Total Score,On Time,Scope Scale,Planning Presentation,Presentation,Idea,Use Cases,Acceptance Criteria,Team Execution,Code,Unit Tests,Repo,Remarks,Repository
+,,,,,100,11,10,6,12,5,4,8,5,23,11,5,,
+1,Leonhard Euler,Daniel Bernoulli,Isaac Newton,Srinivas Ramanujan,92,8.5,8,5,10.5,5,4,8,5,23,10,5,Presentation long and detailed.  Project excellent overall. Need to actually run UI myself.,https://github.com/youngbai/CSYE7200-MovieRecommendation,,,
+2,Ringo Starr,George Harrison,Paul McCartney,,83.5,8.5,10,5,9,5,4,6,3,22,7,4,Presentation quite confused. Part of project not achieved. ,https://github.com/PhoenixMay/MovieRecommendation,,,
+3,Pablo Picasso,Joan Miro,,,92,11,10,5,10,5,4,7,4,23,9,4,Presentation OK. What did they do about these challenges.,https://github.com/jianghanw/CSYE7200_Team_Project_Team3,,,
+4,Ludwig van Beethoven,Wolfgang Amadeus Mozart,,,88.5,11,10,6,10.5,5,3.5,6,4,21.5,7,4,A/C not clearly defined (planning pres?),https://github.com/Essexwwz/heart-health-indicator,,,
+5,Winston Churchill,Clement Attlee,Harold McMillan,,88.5,11,9.5,5,11,5,4,7,2,22,7,5,Presentation long and detailed.  Project excellent overall. A/C not shown clearly.,https://github.com/CSYE7200-21FALL-TEAM6,,,
