+++ cats/src/test/scala/com/phasmidsoftware/tableparser/cats/table/MovieSpec.scala
@@ -6,6 +6,7 @@ package com.phasmidsoftware.tableparser.cats.table
 
 import cats.effect.IO
 import com.phasmidsoftware.tableparser.cats.util.EvaluateIO.{check, checkFailure, matchIO}
+import com.phasmidsoftware.tableparser.core.examples.Movie
 import com.phasmidsoftware.tableparser.core.parse.{CellParser, InvalidParseException, RowParser, StringTableParser}
 import com.phasmidsoftware.tableparser.core.table._
 import org.scalatest.flatspec
@@ -21,7 +22,7 @@ class MovieSpec extends flatspec.AnyFlatSpec with should.Matchers {
   behavior of "Movie table"
 
   it should "parse the first movie from the IMDB dataset" in {
-    import MovieParser._
+    import com.phasmidsoftware.tableparser.core.examples.MovieParser._
 
     val movies = Seq(
       movieHeader,
@@ -39,7 +40,7 @@ class MovieSpec extends flatspec.AnyFlatSpec with should.Matchers {
 
   // CONSIDER rework this test to be more significant
   it should "parse the first (edited) movie from the IMDB dataset" in {
-    import MovieParser._
+    import com.phasmidsoftware.tableparser.core.examples.MovieParser._
 
     val movies = Seq(
       movieHeader,
@@ -53,7 +54,7 @@ class MovieSpec extends flatspec.AnyFlatSpec with should.Matchers {
   }
 
   it should "fail to parse the first (edited) movie from the IMDB dataset" in {
-    import MovieParser._
+    import com.phasmidsoftware.tableparser.core.examples.MovieParser._
 
     implicit object MovieTableParser extends StringTableParser[Table[Movie]] {
       type Row = Movie
@@ -81,7 +82,7 @@ class MovieSpec extends flatspec.AnyFlatSpec with should.Matchers {
   }
 
   it should "parse all the following rows" in {
-    import MovieParser._
+    import com.phasmidsoftware.tableparser.core.examples.MovieParser._
 
     implicit object MovieTableParser extends StringTableParser[Table[Movie]] {
       type Row = Movie
@@ -108,7 +109,7 @@ class MovieSpec extends flatspec.AnyFlatSpec with should.Matchers {
   }
 
   it should "parse and transform the following rows with simple map" in {
-    import MovieParser._
+    import com.phasmidsoftware.tableparser.core.examples.MovieParser._
 
     implicit object MovieTableParser extends StringTableParser[Table[Movie]] {
       type Row = Movie
@@ -142,9 +143,9 @@ class MovieSpec extends flatspec.AnyFlatSpec with should.Matchers {
 
   behavior of "Name"
   it should "parse Philip Michael Thomas" in {
-    import MovieParser._
-    implicitly[CellParser[Name]].convertString("Philip Thomas") shouldBe Success(Name("Philip", None, "Thomas", None))
-    implicitly[CellParser[Name]].convertString("Philip Michael Thomas") shouldBe Success(Name("Philip", Some("Michael"), "Thomas", None))
+    import com.phasmidsoftware.tableparser.core.examples.MovieParser._
+    implicitly[CellParser[com.phasmidsoftware.tableparser.core.examples.Name]].convertString("Philip Thomas") shouldBe Success(com.phasmidsoftware.tableparser.core.examples.Name("Philip", None, "Thomas", None))
+    implicitly[CellParser[com.phasmidsoftware.tableparser.core.examples.Name]].convertString("Philip Michael Thomas") shouldBe Success(com.phasmidsoftware.tableparser.core.examples.Name("Philip", Some("Michael"), "Thomas", None))
 
   }
