+++ core/src/test/scala/com/phasmidsoftware/tableparser/core/table/TableSpec.scala
@@ -4,6 +4,7 @@
 
 package com.phasmidsoftware.tableparser.core.table
 
+import com.phasmidsoftware.tableparser.core.examples.Movie
 import com.phasmidsoftware.tableparser.core.parse._
 import com.phasmidsoftware.tableparser.core.render._
 import com.phasmidsoftware.tableparser.core.table.Table.parseResource
@@ -76,7 +77,7 @@ class TableSpec extends flatspec.AnyFlatSpec with should.Matchers {
   }
 
   it should "parse and not filter the movies from the IMDB dataset" in {
-    import MovieParser._
+    import com.phasmidsoftware.tableparser.core.examples.MovieParser._
     implicit val hasKey: HasKey[Movie] = (t: Movie) => t.production.country
     matchTry(parseResource("movie_metadata.csv", classOf[Movie])) {
       case mt@HeadedTable(_, _) =>
